+++ automatminer/automl/adaptors.py
@@ -19,7 +19,7 @@ from automatminer.utils.ml import is_greater_better, \
     regression_or_classification
 from automatminer.utils.log import log_progress, AMM_LOG_PREDICT_STR, \
     AMM_LOG_FIT_STR
-from automatminer.base import AutoMLAdaptor, LoggableMixin
+from automatminer.base import DFMLAdaptor, LoggableMixin
 
 __authors__ = ['Alex Dunn <ardunn@lbl.gov'
                'Alireza Faghaninia <alireza.faghaninia@gmail.com>',
@@ -31,7 +31,7 @@ _classifier_modes = {'classifier', 'classification', 'classify'}
 _regressor_modes = {'regressor', 'regression', 'regress'}
 
 
-class TPOTAdaptor(AutoMLAdaptor, LoggableMixin):
+class TPOTAdaptor(DFMLAdaptor, LoggableMixin):
     """
     A dataframe adaptor for the TPOT classifiers and regressors.
 
@@ -187,11 +187,6 @@ class TPOTAdaptor(AutoMLAdaptor, LoggableMixin):
         self.models = models
         return best_models_and_scores
 
-    @property
-    @check_fitted
-    def _best_pipeline(self):
-        return self._backend.fitted_pipeline_
-
     @log_progress(AMM_LOG_PREDICT_STR)
     @check_fitted
     def predict(self, df, target):
@@ -232,8 +227,28 @@ class TPOTAdaptor(AutoMLAdaptor, LoggableMixin):
             self.logger.info("Prediction finished successfully.")
             return df
 
+    @property
+    @check_fitted
+    def best_pipeline(self):
+        return self._backend.fitted_pipeline_
 
-class SinglePipelineAdaptor(AutoMLAdaptor, LoggableMixin):
+    @property
+    @check_fitted
+    def features(self):
+        return self._features
+
+    @property
+    @check_fitted
+    def ml_data(self):
+        return self._ml_data
+
+    @property
+    @check_fitted
+    def backend(self):
+        return self._backend
+
+
+class SinglePipelineAdaptor(DFMLAdaptor, LoggableMixin):
     """
     For running single models or pipelines in a MatPipe pipeline using the same
     syntax as the AutoML adaptors.
@@ -312,8 +327,25 @@ class SinglePipelineAdaptor(AutoMLAdaptor, LoggableMixin):
 
     @property
     @check_fitted
-    def _best_pipeline(self):
+    def best_pipeline(self):
         return self._backend
+
+    @property
+    @check_fitted
+    def features(self):
+        return self._features
+
+    @property
+    @check_fitted
+    def ml_data(self):
+        return self._ml_data
+
+    @property
+    @check_fitted
+    def backend(self):
+        return self._backend
+
+
 
 # if __name__ == "__main__":
 #     from matminer.datasets.dataset_retrieval import load_dataset